# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "チーム1のテックブログシステムのためのAPI仕様書です。\n記事の取得、閲覧などの機能を提供します。",
        "title": "Team1 Blog API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/articles": {
            "get": {
                "description": "公開されているブログ記事の一覧を取得します。ログイン済みの場合は内部公開記事も含まれます。ページネーション、部署フィルタ、ステータスフィルタをサポートしています。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "記事 (Articles)"
                ],
                "summary": "記事一覧を取得",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "ページ番号 (デフォルト: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "1ページあたりの件数 (デフォルト: 10, 最大: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "Dev",
                            "MKT",
                            "Ops"
                        ],
                        "type": "string",
                        "description": "部署でフィルタ (Dev, MKT, Ops)",
                        "name": "department",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "internal",
                            "public",
                            "all"
                        ],
                        "type": "string",
                        "description": "ステータスでフィルタ (internal, public, all)。ゲストの場合は常にpublicのみ",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "記事一覧",
                        "schema": {
                            "$ref": "#/definitions/ArticleListResponse"
                        }
                    },
                    "400": {
                        "description": "リクエストパラメータが不正です",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "サーバー内部でエラーが発生しました",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/articles/{slug}": {
            "get": {
                "description": "指定されたslugのブログ記事の詳細を取得します。内部公開記事の場合はログインが必要です。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "記事 (Articles)"
                ],
                "summary": "記事詳細を取得",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"go-api-development\"",
                        "description": "記事のスラグ",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "記事詳細",
                        "schema": {
                            "$ref": "#/definitions/ArticleResponse"
                        }
                    },
                    "403": {
                        "description": "内部公開記事にアクセスするにはログインが必要です",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "記事が見つかりません",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "サーバー内部でエラーが発生しました",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "description": "既存のユーザーを認証し、新しい認証トークンを発行します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "認証 (Auth)"
                ],
                "summary": "ログイン (Log In)",
                "parameters": [
                    {
                        "description": "ユーザー情報 (メールアドレスとパスワード)",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AuthenticateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "認証成功。新しい認証トークンを返します。",
                        "schema": {
                            "$ref": "#/definitions/AuthResponse"
                        }
                    },
                    "400": {
                        "description": "リクエストボディが不正です",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "認証に失敗しました (メールアドレスまたはパスワードが正しくありません)",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "サーバー内部でエラーが発生しました",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/signup": {
            "post": {
                "description": "新しいユーザーアカウントを作成し、認証トークンとユーザー情報を返します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "認証 (Auth)"
                ],
                "summary": "新規ユーザー登録 (Sign Up)",
                "parameters": [
                    {
                        "description": "ユーザー情報 (メールアドレス、パスワード、名前)",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "登録成功。ユーザー情報と認証トークンを返します。",
                        "schema": {
                            "$ref": "#/definitions/AuthResponse"
                        }
                    },
                    "400": {
                        "description": "リクエストボディが不正です",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "指定されたメールアドレスは既に使用されています",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "サーバー内部でエラーが発生しました",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ArticleListResponse": {
            "type": "object",
            "properties": {
                "articles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ArticleResponse"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total_count": {
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "ArticleResponse": {
            "type": "object",
            "properties": {
                "article_type": {
                    "type": "string",
                    "enum": [
                        "markdown",
                        "external"
                    ],
                    "example": "markdown"
                },
                "author": {
                    "$ref": "#/definitions/AuthorResponse"
                },
                "content": {
                    "type": "string",
                    "example": "記事の本文です..."
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-01-06T12:00:00Z"
                },
                "department": {
                    "type": "string",
                    "enum": [
                        "Dev",
                        "MKT",
                        "Ops"
                    ],
                    "example": "Dev"
                },
                "external_url": {
                    "type": "string",
                    "example": "https://example.com/article"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "slug": {
                    "type": "string",
                    "example": "go-api-development"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "internal",
                        "public"
                    ],
                    "example": "public"
                },
                "thumbnail_url": {
                    "type": "string",
                    "example": "https://example.com/thumbnail.jpg"
                },
                "title": {
                    "type": "string",
                    "example": "Go言語でのAPI開発入門"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-01-06T12:00:00Z"
                }
            }
        },
        "AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "$ref": "#/definitions/UserResponse"
                }
            }
        },
        "AuthenticateRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "password123"
                }
            }
        },
        "AuthorResponse": {
            "type": "object",
            "properties": {
                "affiliation": {
                    "type": "string",
                    "example": "開発部"
                },
                "icon_url": {
                    "type": "string",
                    "example": "https://example.com/icon.jpg"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "山田太郎"
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "エラーメッセージ"
                },
                "message": {
                    "type": "string",
                    "example": "詳細なエラー情報"
                }
            }
        },
        "SignUpRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "山田太郎"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "password123"
                }
            }
        },
        "UserResponse": {
            "type": "object",
            "properties": {
                "affiliation": {
                    "type": "string",
                    "example": "開発部"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "icon_url": {
                    "type": "string",
                    "example": "https://example.com/icon.jpg"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "山田太郎"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "認証トークンを'Bearer 'に続けて入力してください。 (例: Bearer {JWTトークン})",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}